+++ CHANGELOG.rst
@@ -10,9 +10,11 @@ Changelog
 3.1.0 - unreleased
 ------------------
 
-**New feature:**
+**New features:**
 
 - New argument ``max_inner_iter`` for classes :class:`~glum.GeneralizedLinearRegressor` and :class:`~glum.GeneralizedLinearRegressorCV` to control the maximum number of iterations of the inner solver in the IRLS-CD algorithm.
+- New fitted attributes ``col_means_`` and ``col_stds_`` for classes :class:`~glum.GeneralizedLinearRegressor` and :class:`~glum.GeneralizedLinearRegressorCV`.
+- :class:`~glum.GeneralizedLinearRegressor` now prints more informative logs when fitting with ``alpha_search=True`` and ``verbose=True``.
 
 **Bug fix:
 
@@ -113,6 +115,10 @@ Changelog
 
 2.5.1 - 2023-05-19
 ------------------
+
+**Other changes:**
+
+- Better error message when the number of input features is different between the fit and predict methods.
 
 **Bug fix:**